requirements.txt
uvicorn
fastapi[all]
langchain
beautifulsoup4
langchain-openai
requests
lxml